Added negative tests

+++ server/resolvers.ts
@@ -3,16 +3,16 @@ import { signToken, hashPassword, verifyPassword } from './utils/index';
 export const resolvers = {
   Query: {
     async users(root, args, { models }) {
-      return models.User.findAll();
+      return await models.User.findAll();
     },
     async userById(root, { id }, { models }) {
-      return models.User.findByPk(id);
+      return await models.User.findByPk(id);
     },
     async courses(root, args, { models }) {
-      return models.Course.findAll();
+      return await models.Course.findAll();
     },
     async courseById(root, { id }, { models }) {
-      return models.Course.findByPk(id);
+      return await models.Course.findByPk(id);
     },
   },
   Mutation: {
@@ -25,16 +25,25 @@ export const resolvers = {
       });
     },
     async updateUser(root, { id, name, email }, { models }) {
-      const user = await models.User.findByPk(id);
-      user.set({
-        name: name,
-        email: email,
-      });
-      return user.save();
+      const user = await models.User.findByPk(id);      
+      if (!user) {
+        return "User not found";
+      } else {
+        user.set({
+          name: name,
+          email: email,
+        });
+        return user.save();
+      }
     },
     async deleteUser(root, { id }, { models }) {
-      await models.User.destroy({ where: { id: id } });
-      return 'User deleted successfully';
+      const user = await models.User.findByPk(id);
+      if (!user) {
+        return "User not found";
+      } else {
+        await models.User.destroy({ where: { id: id } });
+        return 'User deleted successfully';
+      }
     },
 
     async createCourse(root, { userId, name, points }, { models }) {
